Upgrade to omniauth-singly v0.3.0
This puts us onto the new, better `/authenticate` endpoint. The auth URL
generator now auto-appends an `access_token` if it's present in the session so
that when the user is logged in, they connect the profile to their existing
account.

For a full description of the new `/authenticate` endpoint, see
https://singly.com/docs/authorization